Removing testing stuff

+++ src/main/java/nesemulator/utils/ByteUtilities.java
@@ -3,7 +3,7 @@ package nesemulator.utils;
 public class ByteUtilities {
 
     public static int widenIgnoreSigning(byte byteToWiden) {
-        return ((int) byteToWiden & 0xFF);
+        return byteToWiden & 0xFF;
     }
 
     public static int widenIgnoreSigning(short shortToWiden) {
